+++ ase/neb.py
@@ -5,6 +5,7 @@ from math import sqrt
 import numpy as np
 
 import ase.parallel as mpi
+from ase.build import minimize_rotation_and_translation
 from ase.calculators.calculator import Calculator
 from ase.calculators.singlepoint import SinglePointCalculator
 from ase.io import read
@@ -14,8 +15,7 @@ from ase.utils.geometry import find_mic
 
 class NEB:
     def __init__(self, images, k=0.1, climb=False, parallel=False,
-                 world=None, cornercutting=False, improvedtangent=False,
-                 improvedparallelforce=False):
+                 remove_rotation_and_translation=False, world=None):
         """Nudged elastic band.
         Paper I: G. Henkelman and H. Jonsson, Chem. Phys, 113, 9978 (2000).
         Paper II: G. Henkelman, B. P. Uberuaga, and H. Jonsson, Chem. Phys,
@@ -29,19 +29,10 @@ class NEB:
             Use a climbing image (default is no climbing image).
         parallel: bool
             Distribute images over processors.
-        cornercutting: bool
-            If True, springs are treated as real ones with forces in the
-            spring-directions (not the estimated tangent-directions).
-        improvedtangent: bool
-            If False (default), tangents are bisections of spring-directions
-            (formula 2 of paper I).
-            If True, tangents are according to formulas 8, 9,
-            10, and 11 of paper I (end points are bisections since the
-            energies of images 0 and N-1 are not availabe).
-        improvedparallelforce: bool
-            If False (default), original parallel spring force (formula 5 of
-            paper I).
-            If True, improved parallel spring force (formula 12 of paper I).
+        remove_rotation_and_translation: bool
+            TRUE actives NEB-TR for removing translation and
+            rotation during NEB. By default applied non-periodic
+            systems
         """
         self.images = images
         self.climb = climb
@@ -49,10 +40,9 @@ class NEB:
         self.natoms = len(images[0])
         self.nimages = len(images)
         self.emax = np.nan
-        self.cornercutting = cornercutting
-        self.improvedtangent = improvedtangent
-        self.improvedparallelforce = improvedparallelforce
-
+        
+        self.remove_rotation_and_translation = remove_rotation_and_translation
+        
         if isinstance(k, (float, int)):
             k = [k] * (self.nimages - 1)
         self.k = list(k)
@@ -65,8 +55,11 @@ class NEB:
             assert world.size == 1 or world.size % (self.nimages - 2) == 0
 
     def interpolate(self, method='linear', mic=False):
+        if self.remove_rotation_and_translation:
+            minimize_rotation_and_translation(self.images[0], self.images[-1])
+        
         interpolate(self.images, mic)
-
+                 
         if method == 'idpp':
             self.idpp_interpolate(traj=None, log=None, mic=mic)
 
@@ -111,6 +104,12 @@ class NEB:
         images = self.images
         forces = np.empty(((self.nimages - 2), self.natoms, 3))
         energies = np.empty(self.nimages - 2)
+
+        if self.remove_rotation_and_translation:
+            # Remove translation and rotation between
+            # images before computing forces:
+            for i in range(1, self.nimages):
+                minimize_rotation_and_translation(images[i - 1], images[i])
 
         if not self.parallel:
             # Do all images - one at a time: